improved code style
