+++ code/orders-api-rest/src/main/java/com/academy/orders/apirest/cart/controller/CartItemController.java
@@ -1,7 +1,7 @@
-package com.academy.orders.apirest.cart_item.controller;
+package com.academy.orders.apirest.cart.controller;
 
-import com.academy.orders.application.cart_item.usecase.CreateCartItemByUserUseCaseImpl;
-import com.academy.orders.domain.cart_item.entity.CreateCartItemDTO;
+import com.academy.orders.domain.cart.entity.CreateCartItemDTO;
+import com.academy.orders.domain.cart.usecase.CreateCartItemByUserUseCase;
 import com.academy.orders_api_rest.generated.api.CartApi;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -15,11 +15,10 @@ import java.util.UUID;
 @Slf4j
 @CrossOrigin
 public class CartItemController implements CartApi {
-	private final CreateCartItemByUserUseCaseImpl cartItemByUserUseCase;
+	private final CreateCartItemByUserUseCase cartItemByUserUseCase;
 
 	@Override
 	public void addProductToCart(UUID productId, Long userId) {
-		CreateCartItemDTO createCartItemDTO = new CreateCartItemDTO(productId, userId, 1);
-		cartItemByUserUseCase.create(createCartItemDTO);
+		cartItemByUserUseCase.create(new CreateCartItemDTO(productId, userId, 1));
 	}
 }
