+++ code/orders-application/src/main/java/com/academy/orders/application/cart/usecase/CreateCartItemByUserUseCaseUseCaseImpl.java
@@ -1,9 +1,9 @@
-package com.academy.orders.application.cart_item.usecase;
+package com.academy.orders.application.cart.usecase;
 
-import com.academy.orders.domain.cart_item.entity.CreateCartItemDTO;
-import com.academy.orders.domain.cart_item.exception.CartItemAlreadyExistsException;
-import com.academy.orders.domain.cart_item.repository.CartItemRepository;
-import com.academy.orders.domain.cart_item.usecase.CreateCartItemByUser;
+import com.academy.orders.domain.cart.entity.CreateCartItemDTO;
+import com.academy.orders.domain.cart.exception.CartItemAlreadyExistsException;
+import com.academy.orders.domain.cart.repository.CartItemRepository;
+import com.academy.orders.domain.cart.usecase.CreateCartItemByUserUseCase;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -11,7 +11,7 @@ import static java.lang.Boolean.TRUE;
 
 @Service
 @RequiredArgsConstructor
-public class CreateCartItemByUserUseCaseImpl implements CreateCartItemByUser {
+public class CreateCartItemByUserUseCaseUseCaseImpl implements CreateCartItemByUserUseCase {
 	private final CartItemRepository cartItemRepository;
 
 	@Override
