+++ code/orders-domain/src/main/java/com/academy/orders/domain/cart/exception/CartItemAlreadyExistsException.java
@@ -1,13 +1,14 @@
-package com.academy.orders.domain.cart_item.exception;
+package com.academy.orders.domain.cart.exception;
 
 import com.academy.orders.domain.exception.AlreadyExistsException;
 import lombok.Getter;
 
 import java.util.UUID;
 
+@Getter
 public class CartItemAlreadyExistsException extends AlreadyExistsException {
-	@Getter
 	private final UUID productId;
+
 	public CartItemAlreadyExistsException(UUID productId) {
 		super(String.format("Product with %s id already added to the cart", productId));
 		this.productId = productId;
