+++ code/orders-domain/src/main/java/com/academy/orders/domain/cart/usecase/CreateCartItemByUserUseCase.java
@@ -1,0 +1,8 @@
+package com.academy.orders.domain.cart.usecase;
+
+import com.academy.orders.domain.cart.entity.CreateCartItemDTO;
+
+public interface CreateCartItemByUserUseCase {
+
+	void create(CreateCartItemDTO cartItem);
+}
