--- code/orders-domain/src/main/java/com/academy/orders/domain/cart_item/usecase/CreateCartItemByUser.java
@@ -1,8 +1,0 @@
-package com.academy.orders.domain.cart_item.usecase;
-
-import com.academy.orders.domain.cart_item.entity.CreateCartItemDTO;
-
-public interface CreateCartItemByUser {
-
-	void create(CreateCartItemDTO cartItem);
-}
