+++ code/orders-infrastructure/src/main/java/com/academy/orders/infrastructure/cart/repository/CartItemJpaAdapter.java
@@ -1,7 +1,7 @@
-package com.academy.orders.infrastructure.cart_item.repository;
+package com.academy.orders.infrastructure.cart.repository;
 
-import com.academy.orders.infrastructure.cart_item.entity.CartItemEntity;
-import com.academy.orders.infrastructure.cart_item.entity.CartItemId;
+import com.academy.orders.infrastructure.cart.entity.CartItemEntity;
+import com.academy.orders.infrastructure.cart.entity.CartItemId;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
